// src/ignitus-Shared/ignitus-DesignSystem/ignitus-Assets/Ignitus-LazyImage/types.ts
export interface LazyImageTypes {
  alt: string;
  src: string;
  height?: string;
  width?: string;
  width0?: string;
  width1?: string;
  width2?: string;
  width3?: string;
  width4?: string;
  width5?: string;
  height1?: string;
  height2?: string;
  height5?: string;
  padding?: string;
  maxWidth?: string;
  borderRadius?: string;
}